+++ packages/taro-components/src/components/rich-text/index.js
@@ -17,14 +17,19 @@ class RichText extends Nerv.Component {
         style: ''
       }
       if (item.hasOwnProperty('attrs')) {
-        obj.className = item.attrs.class || ''
-        obj.style = item.attrs.style || ''
+        for (const key in item.attrs) {
+          if (key === 'class') {
+            obj.className = item.attrs[key] || ''
+          } else {
+            obj[key] = item.attrs[key] || ''
+          }
+        }
       }
       return Nerv.createElement(item.name, obj, child)
     }
   }
 
-  renderChildrens (arr) {
+  renderChildrens (arr = []) {
     if (arr.length === 0) return
     return arr.map((list, i) => {
       if (list.type === 'text') {
@@ -35,18 +40,18 @@ class RichText extends Nerv.Component {
   }
 
   render () {
-    let { nodes } = this.props
+    let { nodes, className, ...other } = this.props
 
     if (Array.isArray(nodes)) {
       return (
-        <div {...omit(this.props, ['nodes'])}>
+        <div className={className} {...omit(this.props, ['nodes', 'className'])} {...other}>
           {nodes.map((item, idx) => {
             return this.renderNodes(item)
           })}
         </div>
       )
     } else {
-      return <div dangerouslySetInnerHTML={{ __html: nodes }} />
+      return <div className={className} {...omit(this.props, ['className'])} {...other} dangerouslySetInnerHTML={{ __html: nodes }} />
     }
   }
 }