+++ packages/taro-components/src/components/swiper/index.js
@@ -1,7 +1,7 @@
 import 'weui'
 import Nerv from 'nervjs'
 import classNames from 'classnames'
-import * as Swipers from 'swiper'
+import Swipers from 'swiper/dist/js/swiper.min.js'
 
 import 'swiper/dist/css/swiper.min.css'
 import './style/index.scss'
@@ -33,7 +33,8 @@ class Swiper extends Nerv.Component {
       onChange,
       circular,
       vertical,
-      onAnimationfinish
+      onAnimationfinish,
+      spaceBetween
     } = this.props
 
     const opt = {
@@ -76,6 +77,11 @@ class Swiper extends Nerv.Component {
         stopOnLastSlide: true,
         disableOnInteraction: false
       }
+    }
+
+    // 两端距离
+    if (spaceBetween) {
+      opt.spaceBetween = spaceBetween
     }
 
     this.mySwiper = new Swipers(this.$el, opt)