+++ packages/taro-router/src/utils/types.ts
@@ -20,6 +20,7 @@ export interface RouteObj {
   path: Path;
   componentLoader: ComponentLoader;
   isIndex: boolean;
+  isRedirect: boolean;
   key?: string;
 }